refactor: consolidate BENCHMARK_PERSONA and BENCHMARK_TYPE to shared constants
